+++ crates/holochain/src/conductor/api/api_cell.rs
@@ -9,6 +9,7 @@ use crate::core::ribosome::real_ribosome::RealRibosome;
 use crate::core::workflow::ZomeCallResult;
 use async_trait::async_trait;
 use holochain_keystore::MetaLairClient;
+use holochain_p2p::HolochainP2pResult;
 use holochain_state::host_fn_workspace::SourceChainWorkspace;
 use holochain_state::nonce::WitnessNonceResult;
 use holochain_state::prelude::DatabaseResult;
@@ -154,7 +155,7 @@ pub trait CellConductorReadHandleT: Send + Sync {
     ) -> ConductorResult<Option<CellId>>;
 
     /// Expose block functionality to zomes.
-    async fn block(&self, input: Block) -> DatabaseResult<()>;
+    async fn block(&self, input: Block) -> HolochainP2pResult<()>;
 
     /// Expose unblock functionality to zomes.
     async fn unblock(&self, input: Block) -> DatabaseResult<()>;
@@ -256,8 +257,8 @@ impl CellConductorReadHandleT for CellConductorApi {
             .await
     }
 
-    async fn block(&self, input: Block) -> DatabaseResult<()> {
-        self.conductor_handle.block(input).await
+    async fn block(&self, input: Block) -> HolochainP2pResult<()> {
+        self.conductor_handle.holochain_p2p().block(input).await
     }
 
     async fn unblock(&self, input: Block) -> DatabaseResult<()> {