+++ crates/holochain/CHANGELOG.md
@@ -7,6 +7,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+- Remove agents from peer store when they are blocked.
+- Refactor HDK call `block_agent` to use `HolochainP2pActor::block`.
+
 ## 0.6.0-dev.23
 
 - Support binding admin and app websocket interfaces on designated listen addresses. \#5271